Move old db tests to own folder
